Solved bug with setValue in DropDownPicker. Task is created successfully, checked with MongoDB Compass.
The issue was that setValue was giving back a callback function so when we try to do setValue={onChange} with the Controller, we would get a function back. To fix it, we just called the callback function with the value as its argument and called onChange on that. onChange(cb(value)).